feat: restructure settings navigation
- move Your Account to /account, opened from the sidebar user menu; personal API keys become a card there
- add Settings entry to Studio sidebar (last); /settings lands on the first permitted tab
- rename Agents settings tab to Chat; move default MCP-tool guardrails to a new Security tab
- unify API-key vocabulary for personal and service-account credentials; clarify gateway/org token copy with docs links
- cross-link default user limits with Costs & Limits in both directions
- pointer cursor on dropdown menu items; delete confirmation for service-account keys
- update stale docs and in-app references
