+++ docs/pages/platform-quickstart.md
@@ -46,7 +46,7 @@ Once it's up, follow http://localhost:3000
 2. Archestra runs the MCP server in a Kubernetes cluster. To verify, run `kubectl get pods`.
 3. Go to **Agents** and create "Archestra Docs Reader Agent" with this system prompt: `You're using playwright to answer questions about Archestra based on /docs/`
 4. While editing the agent, make sure to enable all `microsoft__playwright-mcp` tools for it.
-5. Go to **Settings -> LLM API Keys** and add your preferred provider. You can connect commercial providers like OpenAI, Anthropic, or Google Gemini. For a free option, use [Cerebras](https://cerebras.ai/) or a local [Ollama](https://ollama.com/) instance.
+5. Go to **Model Providers** (in the Studio sidebar) and add your preferred provider. You can connect commercial providers like OpenAI, Anthropic, or Google Gemini. For a free option, use [Cerebras](https://cerebras.ai/) or a local [Ollama](https://ollama.com/) instance.
 6. Go to **Chat**, choose the "Archestra Docs Reader Agent" and ask "How could I deploy Archestra?"
    ![Archestra Chat UI calling the agent](/docs/quickstart-agent-chat.webp)
    The agent navigates the website to answer the question.
