+++ platform/frontend/src/app/account/_components/change-password-dialog.tsx
