+++ platform/frontend/src/app/account/_components/sessions-card.tsx
