+++ platform/frontend/src/app/connection/mcp-client-instructions.tsx
@@ -450,7 +450,7 @@ function GenericAuthRow({
       <div className="text-xs text-muted-foreground">
         No tokens available — provision one from{" "}
         <Link
-          href="/settings/account?highlight=personal-token"
+          href="/account?highlight=personal-token"
           className="underline hover:text-foreground"
         >
           your account
