+++ platform/frontend/src/app/settings/auth/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function AuthSettingsPage() {
-  redirect("/settings/api-keys");
+  redirect("/account");
 }
