+++ platform/frontend/src/app/settings/settings-tabs.ts
@@ -7,21 +7,23 @@ export function useSettingsTabs() {
   const permissionMap = usePermissionMap(requiredPagePermissionsMap);
   const { data: secretsType } = useSecretsType();
   return [
-    { label: "Your Account", href: "/settings/account" },
-    ...(permissionMap?.["/settings/api-keys"]
-      ? [{ label: "API Keys", href: "/settings/api-keys" }]
+    ...(permissionMap?.["/settings/organization"]
+      ? [{ label: "Organization", href: "/settings/organization" }]
       : []),
     ...(permissionMap?.["/settings/service-accounts"]
       ? [{ label: "Service Accounts", href: "/settings/service-accounts" }]
       : []),
     ...(permissionMap?.["/settings/agents"]
-      ? [{ label: "Agents", href: "/settings/agents" }]
+      ? [{ label: "Chat", href: "/settings/agents" }]
       : []),
     ...(permissionMap?.["/settings/llm"]
       ? [{ label: "LLM", href: "/settings/llm" }]
       : []),
     ...(permissionMap?.["/settings/mcp"]
       ? [{ label: "MCP", href: "/settings/mcp" }]
+      : []),
+    ...(permissionMap?.["/settings/security"]
+      ? [{ label: "Security", href: "/settings/security" }]
       : []),
     ...(permissionMap?.["/settings/knowledge"]
       ? [{ label: "Knowledge", href: "/settings/knowledge" }]
@@ -49,9 +51,6 @@ export function useSettingsTabs() {
       : []),
     ...(secretsType?.type === "Vault" && permissionMap?.["/settings/secrets"]
       ? [{ label: "Secrets", href: "/settings/secrets" }]
-      : []),
-    ...(permissionMap?.["/settings/organization"]
-      ? [{ label: "Organization", href: "/settings/organization" }]
       : []),
   ];
 }
