+++ platform/frontend/src/components/default-credentials-warning.tsx
@@ -110,7 +110,7 @@ export function DefaultCredentialsWarning({
               {" "}
               or{" "}
               <a
-                href="/settings/account?highlight=change-password"
+                href="/account?highlight=change-password"
                 className="inline-flex items-center underline"
               >
                 Change
