+++ platform/frontend/src/components/settings/api-keys-card.utils.ts
